+++ Code/Common/Logger/ResultsLogger.cpp
@@ -1,6 +1,9 @@
 #include "ResultsLogger.h""
 
-// Constructor
+
+
+
+// Constructors
 ResultsLogger::ResultsLogger(const std::string& filename) : m_fileName(filename) {
     outFile.open(filename, std::ios::app); // Append mode
     if (!outFile.is_open()) {
@@ -14,6 +17,54 @@ ResultsLogger::ResultsLogger(const std::string& filename) : m_fileName(filename)
         if (outFile.tellp() == 0) { // If the write pointer is at position 0, the file is empty
             // File is new or empty, write the headers
             outFile << "Test Name,Test ID,Algorithm Name,Degree,Number of Points,Run Time (s),Run Time (us),Number Of Runs,Average Run Time,,Minimum Run Time,TCA Distance,TCA Time\n";
+        }
+        // If the file is not empty, do not write headers
+    }
+}
+
+
+ResultsLogger::ResultsLogger() {
+
+
+    // Get the current time
+    auto now = std::chrono::system_clock::now();
+    auto time_t = std::chrono::system_clock::to_time_t(now);
+    struct tm timeinfo;
+
+    // Convert to local time
+#ifdef _WIN32
+// Use localtime_s on Windows
+    localtime_s(&timeinfo, &time_t);
+#else
+// Use localtime_r on Linux and other UNIX-like systems
+    localtime_r(&time_t, &timeinfo);
+#endif
+
+    // Format the time into a string with the desired format
+    std::ostringstream oss;
+    oss << std::put_time(&timeinfo, "%Y-%m-%d_%H-%M-%S");
+    std::string timeStr = oss.str();
+
+    // Construct the filename using the date and time string
+    std::string filename = "Logger/logger_" + timeStr + ".csv";
+
+    std::cout << "Logger filename: " << filename << std::endl;
+
+
+    m_fileName = filename;
+
+    outFile.open(filename, std::ios::app); // Append mode
+    if (!outFile.is_open()) {
+        std::cout << "Failed to open " << filename << " for writing." << std::endl;
+    }
+    else {
+        std::cout << "Opened " << filename << " for writing." << std::endl;
+
+        // Check if the file is new or empty by checking its size
+        outFile.seekp(0, std::ios::end); // Move the write pointer to the end of the file
+        if (outFile.tellp() == 0) { // If the write pointer is at position 0, the file is empty
+            // File is new or empty, write the headers
+            outFile << "Test Name,Test ID,Algorithm Name,Degree,Number of Points,Run Time (s),Run Time (us),Number Of Runs,Average Run Time,Minimum Run Time,TCA Distance,TCA Time\n";
         }
         // If the file is not empty, do not write headers
     }